requirements.txt
google.generativeai
json
argparse